Agent query: Could you please try saving mood data again from different Telegram accounts? It should work correctly now.
Change user_id column type to BigInteger in MoodEntry model.

Screenshot: https://storage.googleapis.com/screenshot-production-us-central1/105cb7fc-dfd7-4727-b684-658acb2195f0/88fec939-df4f-4abf-9c80-fb2e94a35313.jpg

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Date, JSON, create_engine
+from sqlalchemy import Column, Integer, BigInteger, String, Date, JSON, create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 import os
@@ -14,7 +14,7 @@ class MoodEntry(Base):
     __tablename__ = 'mood_entries'
 
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, nullable=False)
+    user_id = Column(BigInteger, nullable=False)
     date = Column(Date, nullable=False)
     mood = Column(String, nullable=True)
     mood_description = Column(String, nullable=True)
